+++ compiler/src/main/kotlin/dev/zacsweers/metro/compiler/ir/MetroIrGenerationExtension.kt
@@ -6,7 +6,7 @@ import dev.zacsweers.metro.compiler.ClassIds
 import dev.zacsweers.metro.compiler.ExitProcessingException
 import dev.zacsweers.metro.compiler.MetroOptions
 import dev.zacsweers.metro.compiler.Symbols
-import dev.zacsweers.metro.compiler.ir.transformers.ContributionBindsFunctionsIrTransformer
+import dev.zacsweers.metro.compiler.ir.transformers.ContributionTransformer
 import dev.zacsweers.metro.compiler.ir.transformers.DependencyGraphTransformer
 import dev.zacsweers.metro.compiler.ir.transformers.HintGenerator
 import dev.zacsweers.metro.compiler.tracing.trace
@@ -46,15 +46,13 @@ public class MetroIrGenerationExtension(
     try {
       tracer(moduleFragment.name.asString().removePrefix("<").removeSuffix(">"), "Metro compiler")
         .trace { tracer ->
-          // First part 1 - transform $$MetroContribution interfaces to add their binds functions
-          tracer.traceNested("Transforming Metro contributions") {
-            moduleFragment.transform(ContributionBindsFunctionsIrTransformer(context), null)
-          }
-
-          // First part 2 - collect all the contributions in this round
+          // Create contribution data container
           val contributionData = IrContributionData(context)
-          tracer.traceNested("Collecting contributions") {
-            moduleFragment.accept(IrContributionVisitor(context), contributionData)
+
+          // First - transform $$MetroContribution interfaces and collect contribution data in a
+          // single pass
+          tracer.traceNested("Transform contributions") {
+            moduleFragment.transform(ContributionTransformer(context), contributionData)
           }
 
           // Second - transform the dependency graphs